requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile '.\requirements.in'
#
aiohttp==3.8.3
    # via
    #   -r .\requirements.in
    #   azure-functions-durable
aiosignal==1.2.0
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via aiohttp
azure-core==1.25.1
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msrest
    #   opencensus-ext-azure
azure-functions==1.11.2
    # via
    #   -r .\requirements.in
    #   azure-functions-durable
    #   opencensus-extension-azure-functions
azure-functions-durable==1.1.6
    # via -r .\requirements.in
azure-identity==1.11.0
    # via opencensus-ext-azure
azure-storage-blob==12.13.1
    # via -r .\requirements.in
cachetools==5.2.0
    # via google-auth
certifi==2022.9.24
    # via
    #   msrest
    #   requests
    #   trafilatura
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   htmldate
    #   pdfminer-six
    #   requests
    #   trafilatura
courlan==0.8.3
    # via trafilatura
cryptography==38.0.1
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pdfminer-six
    #   pyjwt
dateparser==1.1.1
    # via htmldate
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
furl==2.1.3
    # via azure-functions-durable
google-api-core==2.10.1
    # via opencensus
google-auth==2.12.0
    # via google-api-core
googleapis-common-protos==1.56.4
    # via google-api-core
htmldate==1.3.1
    # via trafilatura
idna==3.4
    # via
    #   requests
    #   yarl
isodate==0.6.1
    # via msrest
justext==3.0.0
    # via trafilatura
langcodes==3.3.0
    # via courlan
lxml==4.9.1
    # via
    #   htmldate
    #   justext
    #   trafilatura
msal==1.19.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.0.0
    # via azure-identity
msrest==0.7.1
    # via azure-storage-blob
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
oauthlib==3.2.1
    # via requests-oauthlib
opencensus==0.11.0
    # via opencensus-ext-azure
opencensus-context==0.1.3
    # via opencensus
opencensus-ext-azure==1.1.7
    # via opencensus-extension-azure-functions
opencensus-extension-azure-functions==1.0.0
    # via -r .\requirements.in
orderedmultidict==1.0.1
    # via furl
pdfminer-six==20220524
    # via -r .\requirements.in
portalocker==2.5.1
    # via msal-extensions
protobuf==4.21.7
    # via
    #   google-api-core
    #   googleapis-common-protos
psutil==5.9.2
    # via opencensus-ext-azure
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.5.0
    # via msal
python-dateutil==2.8.2
    # via
    #   -r .\requirements.in
    #   azure-functions-durable
    #   dateparser
    #   htmldate
pytz==2022.4
    # via dateparser
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pywin32==304 ; sys_platform == "win32"
    # via
    #   -r .\requirements.in
    #   portalocker
regex==2022.3.2
    # via dateparser
requests==2.28.1
    # via
    #   azure-core
    #   azure-functions-durable
    #   google-api-core
    #   msal
    #   msrest
    #   opencensus-ext-azure
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via msrest
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   azure-core
    #   azure-identity
    #   furl
    #   google-auth
    #   isodate
    #   orderedmultidict
    #   python-dateutil
tld==0.12.6
    # via courlan
trafilatura==1.3.0
    # via -r .\requirements.in
types-cryptography==3.3.23
    # via pyjwt
typing-extensions==4.3.0
    # via azure-core
tzdata==2022.4
    # via
    #   pytz-deprecation-shim
    #   tzlocal
tzlocal==4.2
    # via dateparser
urllib3==1.26.12
    # via
    #   courlan
    #   htmldate
    #   requests
    #   trafilatura
yarl==1.8.1
    # via aiohttp
